Add #in_sequence
As explained in the readme, this is inspired by the `do` notation in
Haskell and _sequence comprehensions_ or _for comprehensions_ in Scala.
As also explained in the readme, this utility allows writing code that
looks more like standard, imperative Ruby, while still providing all the
benefits of the Result monad.

As for naming, I chose `in_sequence` because `in_sequence do` reads well
and because both `do` and `for` are reserved keywords in Ruby.
`and_then` and `and_yield` were chosen because `then` and `yield`,
respectively, are also reserved keywords in Ruby. `observe` is similar
in essence to `Result#pipe` already defined within this project because
the result of the function is ignored in both cases, but unlike with
`#pipe`, the function passed to `#observe` only gets invoked if the
chain has been successful this far.